requirements.txt
fastapi
aiosmtplib
python-dotenv
uvicorn
pandas
python-docx
kafka-python
jinja2
